+++ central/globaldb/export/backup.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/stackrox/rox/central/globaldb/v2backuprestore/backup/generators"
+	"github.com/stackrox/rox/central/globaldb/v2backuprestore/backup/generators/cas"
 	"github.com/stackrox/rox/central/globaldb/v2backuprestore/backup/generators/dbs"
 	"github.com/stackrox/rox/pkg/rocksdb"
 	"github.com/stackrox/rox/pkg/utils"
@@ -14,7 +15,7 @@ import (
 )
 
 // Backup backs up the given databases (optionally removing secrets) and writes a ZIP archive to the given writer.
-func Backup(ctx context.Context, boltDB *bolt.DB, rocksDB *rocksdb.RocksDB, out io.Writer) error {
+func Backup(ctx context.Context, boltDB *bolt.DB, rocksDB *rocksdb.RocksDB, includeCerts bool, out io.Writer) error {
 	zipWriter := zip.NewWriter(out)
 	defer utils.IgnoreError(zipWriter.Close)
 
@@ -24,6 +25,12 @@ func Backup(ctx context.Context, boltDB *bolt.DB, rocksDB *rocksdb.RocksDB, out 
 
 	if err := generators.PutTarInZip(generators.PutDirectoryInTar(dbs.NewRocksBackup(rocksDB)), rocksFileName).WriteTo(ctx, zipWriter); err != nil {
 		return errors.Wrap(err, "backing up badger")
+	}
+
+	if includeCerts {
+		if err := generators.PutPathMapInZip(cas.NewCertsBackup(), keysBaseFolder).WriteTo(ctx, zipWriter); err != nil {
+			return errors.Wrap(err, "backing up certificates")
+		}
 	}
 
 	return zipWriter.Close()