+++ central/globaldb/export/file_names.go
@@ -4,4 +4,5 @@ const (
 	boltFileName   = "bolt.db"
 	badgerFileName = "badger.db"
 	rocksFileName  = "rocks.db"
+	keysBaseFolder = "keys"
 )