+++ central/globaldb/v2backuprestore/backup/generators/stream.go
@@ -4,6 +4,7 @@ import (
 	"archive/zip"
 	"context"
 	"io"
+	"os"
 
 	"github.com/pkg/errors"
 )
@@ -32,4 +33,27 @@ func (sgen *fromZipToStream) WriteTo(ctx context.Context, writer io.Writer) erro
 		return errors.Wrap(err, "unable to write to zip file")
 	}
 	return zipeWriter.Close()
+}
+
+// PutFileInStream generates stream from file and write to the output writer.
+func PutFileInStream(filePath string) StreamGenerator {
+	return &fromFileToStream{
+		filePath: filePath,
+	}
+}
+
+type fromFileToStream struct {
+	filePath string
+}
+
+func (s *fromFileToStream) WriteTo(ctx context.Context, writer io.Writer) error {
+	file, err := os.Open(s.filePath)
+	if err != nil {
+		return errors.Wrapf(err, "could not open file %s", s.filePath)
+	}
+	_, err = io.Copy(writer, file)
+	if err != nil {
+		return errors.Wrapf(err, "could not copy file %s", s.filePath)
+	}
+	return nil
 }