+++ central/globaldb/v2backuprestore/backup/generators/tar.go
@@ -17,17 +17,17 @@ type TarGenerator interface {
 
 // PutDirectoryInTar calls the input Directory generator on the input temporary data path, and outpus the results to a tar.
 func PutDirectoryInTar(dGen DirectoryGenerator) TarGenerator {
-	return &fromTarToStream{
+	return &fromDirectoryToTarStream{
 		dGen: dGen,
 	}
 }
 
-type fromTarToStream struct {
+type fromDirectoryToTarStream struct {
 	dGen DirectoryGenerator
 }
 
-func (tGen *fromTarToStream) WriteTo(ctx context.Context, writer *tar.Writer) error {
-	path, err := tGen.dGen.WriteDirectory(ctx)
+func (t *fromDirectoryToTarStream) WriteTo(ctx context.Context, writer *tar.Writer) error {
+	path, err := t.dGen.WriteDirectory(ctx)
 	if err != nil {
 		return errors.Wrap(err, "unable to write to directory")
 	}
@@ -38,4 +38,26 @@ func (tGen *fromTarToStream) WriteTo(ctx context.Context, writer *tar.Writer) er
 	}
 
 	return os.RemoveAll(path)
+}
+
+type fromPathsToTarStream struct {
+	pmGen PathMapGenerator
+}
+
+// PutPathsInTar creates tar from a map of structured paths in tar to its source path.
+func PutPathsInTar(pmGen PathMapGenerator) TarGenerator {
+	return &fromPathsToTarStream{
+		pmGen: pmGen,
+	}
+}
+
+func (t *fromPathsToTarStream) WriteTo(ctx context.Context, writer *tar.Writer) error {
+	pathMap, err := t.pmGen.GeneratePathMap(ctx)
+	if err != nil {
+		return errors.Wrap(err, "unable to generate path map")
+	}
+	if err = pkgTar.FromPathMap(pathMap, writer); err != nil {
+		return errors.Wrap(err, "unable to tar path map")
+	}
+	return nil
 }