+++ central/globaldb/v2backuprestore/backup/generators/zip.go
@@ -4,6 +4,7 @@ import (
 	"archive/tar"
 	"archive/zip"
 	"context"
+	"path"
 
 	"github.com/pkg/errors"
 )
@@ -68,6 +69,34 @@ func (zgen *fromTarToZip) WriteTo(ctx context.Context, writer *zip.Writer) error
 	err = tarWriter.Close()
 	if err != nil {
 		return errors.Wrapf(err, "unable to close %s to zip", zgen.fileName)
+	}
+	return nil
+}
+
+type fromPathMapToZip struct {
+	pmGen    PathMapGenerator
+	basePath string
+}
+
+// PutPathMapInZip gets the destination and source path from the path map and zips the files to destination in zip.
+func PutPathMapInZip(pmGen PathMapGenerator, basePath string) ZipGenerator {
+	return &fromPathMapToZip{
+		pmGen:    pmGen,
+		basePath: basePath,
+	}
+}
+
+func (z *fromPathMapToZip) WriteTo(ctx context.Context, writer *zip.Writer) error {
+	pathMap, err := z.pmGen.GeneratePathMap(ctx)
+	if err != nil {
+		return errors.Wrap(err, "unable to generate path map")
+	}
+	for toPath, fromPath := range pathMap {
+		stream := PutFileInStream(fromPath)
+		err := PutStreamInZip(stream, path.Join(z.basePath, toPath)).WriteTo(ctx, writer)
+		if err != nil {
+			return errors.Wrapf(err, "error creating %s in zip", toPath)
+		}
 	}
 	return nil
 }