+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Entries in this file should be limited to:
 Please avoid adding duplicate information across this changelog and JIRA/doc input pages.
 
 ## [NEXT RELEASE]
+- Added option to backup certificates for central.
 
 ## [53.0]
 - No Changes to note